完成了玩家状态机的设置。并修改了玩家奔跑动画未循环播放的问题；并修改玩家攻击方式，从鼠标点击攻击修改为键盘操控。

+++ character.cpp
@@ -3,9 +3,6 @@
 //
 
 #include "character.h"
-
-#include <iso646.h>
-
 #include "collision_manager.h"
 
 Character::Character() {
@@ -55,7 +52,7 @@ void Character::on_update(float delta) {
     }
 
     if (position.x <= 0) position.x = 0;
-    if (position.x >= getwidth()) position.x = static_cast<float>(getwidth());
+    if (position.x >= static_cast<float>(getwidth())) position.x = static_cast<float>(getwidth());
 
     hurt_box->set_position(get_logic_center());
